Moved reducers and actions belonging to the queue data store from SampleGrid to queue.
Some unnecessary auxiliary functions (addTask and removeTask) were
removed and merged with their corresponding callees.

sendDeleteSampleTask was changed so that it takes taskData and queueID.
The functionality of the function and its corresponding reducer was not
working properly (due to previous incomplete merge ? ).
