+++ mxcube3/ui/actions/queue.js
@@ -1,8 +1,95 @@
 import fetch from 'isomorphic-fetch';
+import { setLoading, showErrorPanel } from './general';
+import { showTaskForm } from './taskForm';
 
-export function addSample(sampleID, queueID) {
+
+export function setSampleListAction(sampleList) {
+  return { type: 'SET_SAMPLE_LIST', sampleList };
+}
+
+
+export function sendGetSampleListRequest() {
+  return function (dispatch) {
+    dispatch(setLoading(true));
+    fetch('mxcube/api/v0.1/sample_changer/samples_list', { credentials: 'include' })
+                        .then(response => response.json())
+                        .then(json => {
+                          dispatch(setLoading(false));
+                          dispatch(setSampleListAction(json));
+                        }, () => {
+                          dispatch(setLoading(false));
+                          dispatch(showErrorPanel(true, 'Could not get samples list'));
+                        });
+  };
+}
+
+
+export function clearAll() {
   return {
-    type: 'ADD_SAMPLE', sampleID, queueID
+    type: 'CLEAR_ALL'
+  };
+}
+
+
+export function sendClearQueue() {
+  return function (dispatch) {
+    fetch('mxcube/api/v0.1/queue/clear', {
+      method: 'PUT',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      }
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Server refused to clear queue');
+      }
+    }).then(() => {
+      dispatch(clearAll());
+    });
+  };
+}
+
+
+export function setManualMountAction(manual) {
+  return { type: 'SET_MANUAL_MOUNT', manual };
+}
+
+
+export function sendManualMountRequest(manual) {
+  return function (dispatch) {
+    return fetch('mxcube/api/v0.1/diffractometer/usesc', {
+      method: 'PUT',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      },
+      body: JSON.stringify({ use_sc: !manual })
+    }).then((response) => {
+      if (response.status >= 400) {
+        dispatch(showErrorPanel(true, 'Could not toogle manual mode'));
+      } else {
+        dispatch(sendClearQueue());
+        dispatch(setSampleListAction({}));
+        dispatch(setManualMountAction(manual));
+        if (manual) {
+          dispatch(showTaskForm('AddSample'));
+        }
+      }
+    });
+  };
+}
+
+
+export function setSampleOrderAction(newSampleOrder, keys) {
+  return { type: 'SET_SAMPLE_ORDER', order: newSampleOrder, keys };
+}
+
+
+export function addSample(sampleID, queueID, sampleData) {
+  return {
+    type: 'ADD_SAMPLE', sampleID, queueID, sampleData
   };
 }
 
@@ -12,11 +99,6 @@ export function removeSample(queueID, sampleID) {
   };
 }
 
-export function clearAll() {
-  return {
-    type: 'CLEAR_ALL'
-  };
-}
 
 export function setStatus(queueState) {
   return {
@@ -166,66 +248,6 @@ export function sendStopQueue() {
   };
 }
 
-export function sendClearQueue() {
-  return function (dispatch) {
-    fetch('mxcube/api/v0.1/queue/clear', {
-      method: 'PUT',
-      credentials: 'include',
-      headers: {
-        Accept: 'application/json',
-        'Content-type': 'application/json'
-      }
-    }).then((response) => {
-      if (response.status >= 400) {
-        throw new Error('Server refused to clear queue');
-      }
-    }).then(() => {
-      dispatch(clearAll());
-    });
-  };
-}
-
-export function sendAddSample(SampleId) {
-  return function (dispatch) {
-    return fetch('mxcube/api/v0.1/queue', {
-      method: 'POST',
-      credentials: 'include',
-      headers: {
-        Accept: 'application/json',
-        'Content-type': 'application/json'
-      },
-      body: JSON.stringify({ SampleId })
-    }).then((response) => {
-      if (response.status >= 400) {
-        throw new Error('Server refused to add sample to queue');
-      }
-      return response.json();
-    }).then((json) => {
-      dispatch(addSample(json.SampleId, json.QueueId));
-      return json.QueueId; // dispatch(sendState());
-    });
-  };
-}
-
-
-export function sendDeleteSample(queueID, sampleID) {
-  return function (dispatch) {
-    fetch(`mxcube/api/v0.1/queue/${queueID}`, {
-      method: 'DELETE',
-      credentials: 'include',
-      headers: {
-        Accept: 'application/json',
-        'Content-type': 'application/json'
-      }
-    }).then((response) => {
-      if (response.status >= 400) {
-        throw new Error('Server refused to remove sample');
-      } else {
-        dispatch(removeSample(queueID, sampleID));
-      }
-    });
-  };
-}
 
 export function sendMountSample(queueID) {
   return function (dispatch) {
@@ -246,10 +268,35 @@ export function sendMountSample(queueID) {
   };
 }
 
-export function sendUnmountSample(queueID) {
+
+export function sendAddSample(SampleId, sampleData) {
   return function (dispatch) {
-    fetch(`mxcube/api/v0.1/sample_changer/${queueID}/unmount`, {
-      method: 'PUT',
+    return fetch('mxcube/api/v0.1/queue', {
+      method: 'POST',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      },
+      body: JSON.stringify({ SampleId })
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Server refused to add sample to queue');
+      }
+      return response.json();
+    }).then((json) => {
+      dispatch(addSample(json.SampleId, json.QueueId, sampleData));
+      dispatch(sendMountSample(json.QueueId));
+      return json.QueueId; // dispatch(sendState());
+    });
+  };
+}
+
+
+export function sendDeleteSample(queueID, sampleID) {
+  return function (dispatch) {
+    return fetch(`mxcube/api/v0.1/queue/${queueID}`, {
+      method: 'DELETE',
       credentials: 'include',
       headers: {
         Accept: 'application/json',
@@ -257,13 +304,14 @@ export function sendUnmountSample(queueID) {
       }
     }).then((response) => {
       if (response.status >= 400) {
-        throw new Error('Server refused to unmount sample');
+        throw new Error('Server refused to remove sample');
       } else {
-        dispatch(unmountSample(queueID));
+        dispatch(removeSample(queueID, sampleID));
       }
     });
   };
 }
+
 
 export function sendRunSample(queueID) {
   return function (dispatch) {
@@ -279,6 +327,141 @@ export function sendRunSample(queueID) {
         throw new Error('Server refused to run sample');
       } else {
         dispatch(runSample(queueID));
+      }
+    });
+  };
+}
+
+
+export function addTaskAction(sampleQueueID, sampleID, task, parameters) {
+  return { type: 'ADD_TASK',
+           taskType: task.Type,
+           sampleID,
+           parentID: sampleQueueID,
+           queueID: task.QueueId,
+           parameters
+  };
+}
+
+
+export function sendAddSampleTaskRequest(queueID, sampleID, parameters, runNow) {
+  return function (dispatch) {
+    fetch(`mxcube/api/v0.1/queue/${queueID}`, {
+      method: 'POST',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      },
+      body: JSON.stringify(parameters)
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Could not add sample task, server refused');
+      }
+      return response.json();
+    }).then((json) => {
+      if (runNow) {
+        dispatch(sendRunSample(json.QueueId));
+      }
+      dispatch(addTaskAction(queueID, sampleID, json, parameters));
+    });
+  };
+}
+
+
+export function sendAddSampleAndTaskRequest(sampleID, parameters) {
+  return function (dispatch) {
+    dispatch(sendAddSample(sampleID)).then(
+      queueID => {
+        dispatch(sendAddSampleTaskRequest(queueID, sampleID, parameters));
+      });
+  };
+}
+
+
+export function updateTaskAction(queueID, sampleID, parameters) {
+  return { type: 'UPDATE_TASK',
+           index: sampleID,
+           queueID,
+           parameters
+         };
+}
+
+
+export function sendUpdateSampleTaskRequest(taskQueueID, sampleQueueID, sampleID, params, runNow) {
+  return function (dispatch) {
+    fetch(`mxcube/api/v0.1/queue/${sampleQueueID}/${taskQueueID}`, {
+      method: 'PUT',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      },
+      body: JSON.stringify(params)
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Could not change sample task, server refused');
+      }
+      return response.json();
+    }).then(() => {
+      if (runNow) {
+        dispatch(sendRunSample(taskQueueID));
+      }
+      dispatch(updateTaskAction(taskQueueID, sampleID, params));
+    });
+  };
+}
+
+
+export function removeTaskAction(task) {
+  return { type: 'REMOVE_TASK', task };
+}
+
+
+export function sendDeleteSampleTask(task, queueID) {
+  return function (dispatch) {
+    fetch(`mxcube/api/v0.1/queue/${queueID}`, {
+      method: 'DELETE',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      }
+
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Server refused to remove sample');
+      } else {
+        dispatch(removeTaskAction(task));
+      }
+    });
+  };
+}
+
+
+export function addTaskResultAction(sampleID, taskQueueID, state) {
+  return { type: 'ADD_TASK_RESLT',
+           index: sampleID,
+           queueID: taskQueueID,
+           state
+  };
+}
+
+
+export function sendUnmountSample(queueID) {
+  return function (dispatch) {
+    fetch(`mxcube/api/v0.1/sample_changer/${queueID}/unmount`, {
+      method: 'PUT',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json',
+        'Content-type': 'application/json'
+      }
+    }).then((response) => {
+      if (response.status >= 400) {
+        throw new Error('Server refused to unmount sample');
+      } else {
+        dispatch(unmountSample(queueID));
       }
     });
   };
